% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_object_times.R
\name{generate_object_times}
\alias{generate_object_times}
\title{Generate Object Times}
\usage{
generate_object_times(n = 30, distr, time_names, export_file = NA)
}
\arguments{
\item{n}{number of items to generate.}

\item{distr}{A list of info on the distributions used for each time.}

\item{time_names}{A vector of names for the different time values.}

\item{export_file}{Filename to export dataframe in csv format. When this is
set to the default value NA, no csv file will be exported.}
}
\value{
A dataframe of draws of times.
}
\description{
Generate the draws for the inter-arrival time, the process times for the
simulation
}
\details{
Draw different times for all the distributions.
}
\examples{
generate_object_times(n=30, distr=list(c(10,2),c(15,1),c(20,2),c(10,0.5)),
               time_names=c("inter_arrival_time", "process_time1",
               "process_time2", "process_time3"))
}
\author{
Yves R. Sagaert
}
